requirements.txt
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Flask-Login==0.4.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
SQLAlchemy==1.2.13
Werkzeug==0.14.1
WTForms==2.2.1
psycopg2==2.7.6.1
Pillow==5.3.0
